# openapi.yaml
openapi: 3.1.0
info:
  title: Galaxy of Consequence RPG Backend API
  description: A comprehensive Star Wars RPG backend with NVIDIA Nemotron integration, persistent world simulation, reactive NPCs, and immersive Force mechanics following the Final Rule.
  version: 1.0.0
  contact:
    name: Galaxy of Consequence API
    email: api@galaxyofconsequence.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.galaxyofconsequence.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /update_alignment:
    post:
      tags:
        - Force
      summary: Update Force alignment
      description: Update player Force alignment with cascading consequences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user
                - action_type
              properties:
                user:
                  type: string
                action_type:
                  type: string
                  enum: [light, dark, neutral]
                action_description:
                  type: string
                force_magnitude:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 1
                witnesses:
                  type: array
                  items:
                    type: string
                session_id:
                  type: string
      responses:
        '200':
          description: Force alignment updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  alignment_change:
                    type: integer
                  new_alignment:
                    type: integer
                  alignment_description:
                    type: string
                  force_consequences:
                    type: array
                  unlocked_powers:
                    type: array
                  corruption_level:
                    type: integer
                  npc_reactions:
                    type: array
  /get_alignment:
    get:
      tags:
        - Force
      summary: Get Force alignment
      parameters:
        - name: user
          in: query
          required: true
          schema:
            type: string
        - name: session_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Force alignment retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForceAlignment'
  /use_force_power:
    post:
      tags:
        - Force
      summary: Use Force power
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user
                - power_name
              properties:
                user:
                  type: string
                power_name:
                  type: string
                target:
                  type: string
                intent:
                  type: string
                  enum: [light, dark, neutral]
                  default: neutral
                power_level:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 1
                session_id:
                  type: string
      responses:
        '200':
          description: Force power used
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
                  effect_description:
                    type: string
                  alignment_change:
                    type: integer
                  force_cost:
                    type: integer
                  consequences:
                    type: array
                  witnesses_affected:
                    type: array
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Use 'Abracadabra' as the Bearer token
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /authenticate endpoint
  schemas:
...
    ForceAlignment:
      type: object
      properties:
        user:
          type: string
        force_sensitive:
          type: boolean
        light_side_points:
          type: integer
        dark_side_points:
          type: integer
        net_alignment:
          type: integer
        alignment_description:
          type: string
        corruption_level:
          type: integer
        force_events:
          type: array
        alignment_history:
          type: array
        force_powers:
          type: array
        last_force_event:
          type: string
...